Spelling fix
self explanatory requires a hyphen, I'm sure that once this is fixed,
the code will run much smoother as there will be no syntax errors
anywhere in this project. This is very sound logic. I'm definitely
contributing much to this project. Export a lot more of these very
useful commits to come. Oh, and by the way, your welcome.